small refactoring: define constants, change todo

+++ src/main/java/com/example/tedtalksanalyzer/dto/TedTalkAnalyticsDTO.java
@@ -10,7 +10,7 @@ import java.util.Map;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TedTalkAnalyticsDTO {
-    private Map<String, Long> speakerInfluence;
+    private Map<String, Double> speakerInfluence;
     private Map<Integer, String> bestTalkPerYear;
     private Map<Integer, Long> talksPerYear;
 }
